requirement.txt
django
djangorestframework
pillow
django-filter
